Added a hacky implementation of grep-like output when running exq.
This implementation prints to stdout and does not perform error
checks. It is only useful to prove to myself I understand the
data structure of the doc.Locations field.

The next step must be to fully implement the logic to extract
the context around the matched term. The first iteration can be
simple and grep-like, e.g.,

filename: the brown <color>fox</color> jumped over...

Which means it contains the file name, indexes of the surrounding
newlines.

Future iteration can be more intelligent, somehow.