DEV: Change "4gt qc" to do model transformation on client side
Previously, the logic to transform []DocumentMatch to []ResultEntry
was done on the server-side, which tightly couples how things are
rendered to the Atlas API.

By having the Atlas API return a bleve SearchResult object, we can
more easily implement different ways to render the search results.

This sets the stage for implementing a result view that shows
one text snippet per matched document.